Add files via upload
In the context of automotive sales, this study
addresses the imperative for a predictive model to estimate
consumer expenditure based on various customer attributes
such as name, email, country, gender, age, annual salary,
credit card debt, and net worth. The focus of the model is
the anticipation of the amount paid for a car, constituting a
regression task. The comparative analysis explores the
efficacy of three distinct algorithms—Artificial Neural
Networks (ANN), Multiple Linear Regression, Lasso and
Random Forest—in predicting car sales prices. The research
aims to discern the strengths and weaknesses of each
algorithm, shedding light on their respective performances
in the given context. By delving into this integrative
approach, the study seeks to provide valuable insights for
automobile salespersons, offering a nuanced understanding
of predictive modeling techniques for enhanced
decision-making in the dynamic realm of automotive sales.